+++ test/rendering/browser/webgl2-render-to-texture.test.ts
@@ -22,9 +22,7 @@ type RGBATuple = [number, number, number, number];
 
 const canvasSize = 64;
 const defaultWebGlAttributes: WebGLContextAttributes = {
-  alpha: false,
   antialias: false,
-  premultipliedAlpha: false,
   preserveDrawingBuffer: true,
   stencil: false,
   depth: false,